"update ping" commit
Added functionality to the collector so it will check if reporter is
already in the database on a '/register' call.
If the reporter is in the database then it will check if the database
ping value and the ping value passed in through the '/register' call are
different.
If the values are different the application with update the database
ping value, and if the values are the same the application does nothing.

This functionality is to address the scenario where a reporter is
shutdown and restarted in order to update the ping interval time